feat: Added chat feature
Was added a chat feature which allows users to create new chat rooms and exchange messages for each room

+++ BackEnd/JobsityChatChallenge/Chat.Core/AppContext/Repository/IRepository.cs
@@ -15,6 +15,6 @@ namespace Chat.Core.AppContext.Repository
         void Update(T entity);
         IQueryable<T> Get(Expression<Func<T, bool>> filter = null, Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null, params Expression<Func<T, object>>[] includes);
         IQueryable<T> Query(Expression<Func<T, bool>> filter = null, Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
